Changed fabs to fabsf to potentially get back a tiny bit of performance

+++ BasicDiffuseSpecularShader.cpp
@@ -62,10 +62,10 @@ void BasicDiffuseSpecularShader::shade( Scene & scene, RandomNumberGenerator & r
             }
 
             // Not in shadow
-            float cos_r_n = fabs( dot( direction, intersection.normal ) ); 
+            float cos_r_n = fabsf( dot( direction, intersection.normal ) ); 
             RGBColor color = light->material->emittance;
             color.scale( cos_r_n );
-            color.scale( fabs( dot( sample.normal, direction.negated() ) ) ); // account for projected area of the light
+            color.scale( fabsf( dot( sample.normal, direction.negated() ) ) ); // account for projected area of the light
             // FIXME - Should there be a distance falloff? Things seem too bright for objects close to lights if we do
             //color.scale( 1.0f / dist_sq_to_light ); // distance falloff
             // TODO: use actual material parameters properly so we can get specular here, too
@@ -89,7 +89,7 @@ void BasicDiffuseSpecularShader::shade( Scene & scene, RandomNumberGenerator & r
         }
 
         // Not in shadow
-        float cos_r_n = fabs( dot( direction, intersection.normal ) ); 
+        float cos_r_n = fabsf( dot( direction, intersection.normal ) ); 
         RGBColor color = light.band_power;
         color.scale(cos_r_n);
         color.scale(1.0f / to_light.magnitude_sq()); // distance falloff
